requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
websockets==14.1
openai==1.58.1
twilio==9.4.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
jinja2==3.1.5
python-multipart==0.0.20
bcrypt==4.2.1
httpx==0.28.1
